Refactor BaseMod Init out to separate class
